backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
httpx
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
boto3